math16 - cast range checks

+++ libsrc/_DEVELOPMENT/math/float/math16/c/log10f16.c
@@ -57,7 +57,7 @@ half_t log10f16 (half_t x)
 
     /* Test for domain */
     if( x <= 0.0 )
-        return( HUGE_NEG_F16 );
+        return (half_t)HUGE_NEG_F16;
 
     /* separate mantissa from exponent */
     x = frexpf16( x, &e );
